requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
